requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
asyncio==3.4.3
attrs==23.2.0
bardapi==0.1.39
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.2.1
decorator==5.1.1
distro==1.9.0
distutils-pytest==0.2.1
disutils==1.4.32.post2
executing==2.0.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
iniconfig==2.0.0
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
MouseInfo==0.1.3
multidict==6.0.5
numpy==1.26.4
openai==0.28.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
pytest==8.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
pytweening==1.0.7
pytz==2024.1
pywin32==306
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.1
stack-data==0.6.3
streamlit==1.32.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
watchdog==4.0.0
wcwidth==0.2.13
wikipedia==1.4.0
yarl==1.9.4